chore: make setup a lot more simplified
